+++ htmlcov/class_index.html
@@ -11,7 +11,7 @@
 <header>
     <div class="content">
         <h1>Coverage report:
-            <span class="pc_cov">79%</span>
+            <span class="pc_cov">93%</span>
         </h1>
         <aside id="help_panel_wrapper">
             <input id="help_panel_state" type="checkbox">
@@ -55,7 +55,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:45 +0300
+            created at 2025-07-09 13:02 +0300
         </p>
     </div>
 </header>
@@ -145,36 +145,44 @@
                 <td class="right" data-ratio="6 6">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t10">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t10"><data value='Smartphone'>Smartphone</data></a></td>
-                <td>5</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="5 5">100%</td>
-            </tr>
-            <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t38">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t38"><data value='LawnGrass'>LawnGrass</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_lawngrass_py.html#t3">src\lawngrass.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_lawngrass_py.html#t3"><data value='LawnGrass'>LawnGrass</data></a></td>
                 <td>4</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="4 4">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t62">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t62"><data value='Order'>Order</data></a></td>
-                <td>4</td>
+                <td class="name left"><a href="z_145eef247bfb46b6_lawngrass_py.html">src\lawngrass.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_lawngrass_py.html"><data value=''><span class='no-noun'>(no class)</span></data></a></td>
+                <td>6</td>
                 <td>0</td>
                 <td>0</td>
-                <td class="right" data-ratio="4 4">100%</td>
+                <td class="right" data-ratio="6 6">100%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_main_py.html">src\main.py</a></td>
                 <td class="name left"><a href="z_145eef247bfb46b6_main_py.html"><data value=''><span class='no-noun'>(no class)</span></data></a></td>
-                <td>50</td>
-                <td>29</td>
+                <td>6</td>
                 <td>0</td>
-                <td class="right" data-ratio="21 50">42%</td>
+                <td>0</td>
+                <td class="right" data-ratio="6 6">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_order_py.html#t4">src\order.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_order_py.html#t4"><data value='Order'>Order</data></a></td>
+                <td>4</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="4 4">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_order_py.html">src\order.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_order_py.html"><data value=''><span class='no-noun'>(no class)</span></data></a></td>
+                <td>7</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="7 7">100%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_print_mixin_py.html#t4">src\print_mixin.py</a></td>
@@ -209,22 +217,38 @@
                 <td class="right" data-ratio="16 16">100%</td>
             </tr>
             <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_smartphone_py.html#t3">src\smartphone.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_smartphone_py.html#t3"><data value='Smartphone'>Smartphone</data></a></td>
+                <td>5</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="5 5">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_smartphone_py.html">src\smartphone.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_smartphone_py.html"><data value=''><span class='no-noun'>(no class)</span></data></a></td>
+                <td>7</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="7 7">100%</td>
+            </tr>
+            <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_utils_py.html">src\utils.py</a></td>
                 <td class="name left"><a href="z_145eef247bfb46b6_utils_py.html"><data value=''><span class='no-noun'>(no class)</span></data></a></td>
-                <td>18</td>
+                <td>19</td>
                 <td>8</td>
                 <td>0</td>
-                <td class="right" data-ratio="10 18">56%</td>
+                <td class="right" data-ratio="11 19">58%</td>
             </tr>
         </tbody>
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
                 <td class="name left">&nbsp;</td>
-                <td>196</td>
-                <td>41</td>
+                <td>173</td>
+                <td>12</td>
                 <td>0</td>
-                <td class="right" data-ratio="155 196">79%</td>
+                <td class="right" data-ratio="161 173">93%</td>
             </tr>
         </tfoot>
     </table>
@@ -236,7 +260,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:45 +0300
+            created at 2025-07-09 13:02 +0300
         </p>
     </div>
     <aside class="hidden">
