+++ htmlcov/index.html
@@ -11,7 +11,7 @@
 <header>
     <div class="content">
         <h1>Coverage report:
-            <span class="pc_cov">79%</span>
+            <span class="pc_cov">93%</span>
         </h1>
         <aside id="help_panel_wrapper">
             <input id="help_panel_state" type="checkbox">
@@ -54,7 +54,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:45 +0300
+            created at 2025-07-09 13:02 +0300
         </p>
     </div>
 </header>
@@ -106,11 +106,25 @@
                 <td class="right" data-ratio="16 16">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html">src\main.py</a></td>
-                <td>63</td>
-                <td>29</td>
+                <td class="name left"><a href="z_145eef247bfb46b6_lawngrass_py.html">src\lawngrass.py</a></td>
+                <td>10</td>
                 <td>0</td>
-                <td class="right" data-ratio="34 63">54%</td>
+                <td>0</td>
+                <td class="right" data-ratio="10 10">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html">src\main.py</a></td>
+                <td>6</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="6 6">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_order_py.html">src\order.py</a></td>
+                <td>11</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="11 11">100%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_print_mixin_py.html">src\print_mixin.py</a></td>
@@ -127,20 +141,27 @@
                 <td class="right" data-ratio="47 47">100%</td>
             </tr>
             <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_smartphone_py.html">src\smartphone.py</a></td>
+                <td>12</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="12 12">100%</td>
+            </tr>
+            <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_utils_py.html">src\utils.py</a></td>
-                <td>18</td>
+                <td>19</td>
                 <td>8</td>
                 <td>0</td>
-                <td class="right" data-ratio="10 18">56%</td>
+                <td class="right" data-ratio="11 19">58%</td>
             </tr>
         </tbody>
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>196</td>
-                <td>41</td>
+                <td>173</td>
+                <td>12</td>
                 <td>0</td>
-                <td class="right" data-ratio="155 196">79%</td>
+                <td class="right" data-ratio="161 173">93%</td>
             </tr>
         </tfoot>
     </table>
@@ -152,7 +173,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:45 +0300
+            created at 2025-07-09 13:02 +0300
         </p>
     </div>
     <aside class="hidden">
