+++ src/lawngrass.py
@@ -1,0 +1,24 @@
+from src.product import Product
+
+class LawnGrass(Product):
+    """Класс категории газонная трава наследуемый от Product"""
+
+    country: str
+    germination_period: str
+    color: str
+
+    def __init__(
+        self,
+        name: str,
+        description: str,
+        price: float,
+        quantity: int,
+        country: str,
+        germination_period: str,
+        color: str,
+    ):
+        """Инициализация класса, наследник Product"""
+        super().__init__(name, description, price, quantity)
+        self.country = country
+        self.germination_period = germination_period
+        self.color = color
