+++ src/main.py
@@ -1,123 +1,7 @@
-from src.base_category_order import BaseCategoryOrder
 from src.product import Product
 from src.category import Category
 from src.category_iterator import CategoryIterator
+from src.smartphone import Smartphone
+from src.lawngrass import LawnGrass
+from src.order import Order
 
-
-
-
-
-class Smartphone(Product):
-    """Класс категории смартфоны наследуемый от Product"""
-
-    efficiency: float
-    model: str
-    memory: int
-    color: str
-
-    def __init__(
-        self,
-        name: str,
-        description: str,
-        price: float,
-        quantity: int,
-        efficiency: float,
-        model: str,
-        memory: int,
-        color: str,
-    ):
-        """Инициализация класса смартфонов наследник Product"""
-
-        super().__init__(name, description, price, quantity)
-        self.efficiency = efficiency
-        self.model = model
-        self.memory = memory
-        self.color = color
-
-
-class LawnGrass(Product):
-    """Класс категории газонная трава наследуемый от Product"""
-
-    country: str
-    germination_period: str
-    color: str
-
-    def __init__(
-        self,
-        name: str,
-        description: str,
-        price: float,
-        quantity: int,
-        country: str,
-        germination_period: str,
-        color: str,
-    ):
-        """Инициализация класса, наследник Product"""
-        super().__init__(name, description, price, quantity)
-        self.country = country
-        self.germination_period = germination_period
-        self.color = color
-
-
-class Order(BaseCategoryOrder):
-    """Класс заказа одного товара"""
-
-    def __init__(self, product: Product, quantity: int):
-        """Инициализация класса заказа"""
-
-        self.product = product
-        self.quantity = quantity
-
-    @property
-    def total_price(self) -> float:
-        """Возвращает стоимость заказа произведение стоимости и количества продуктов"""
-
-        return self.product.price * self.quantity
-
-    def __str__(self) -> str:
-        """возвращает информацию о заказе"""
-
-        return f"{self.product.name}, количество продуктов: {self.quantity}, общая стоимость: {self.total_price}"
-
-if __name__ == '__main__':
-    product1 = Product("Samsung Galaxy S23 Ultra", "256GB, Серый цвет, 200MP камера", 180000.0, 5)
-    product2 = Product("Iphone 15", "512GB, Gray space", 210000.0, 8)
-    product3 = Product("Xiaomi Redmi Note 11", "1024GB, Синий", 31000.0, 14)
-
-    print(product1.name)
-    print(product1.description)
-    print(product1.price)
-    print(product1.quantity)
-
-    print(product2.name)
-    print(product2.description)
-    print(product2.price)
-    print(product2.quantity)
-
-    print(product3.name)
-    print(product3.description)
-    print(product3.price)
-    print(product3.quantity)
-
-    category1 = Category("Смартфоны",
-                         "Смартфоны, как средство не только коммуникации, но и получения дополнительных функций для удобства жизни",
-                         [product1, product2, product3])
-
-    print(category1.name == "Смартфоны")
-    print(category1.description)
-    print(len(category1.products))
-    print(category1.category_count)
-    print(category1.product_count)
-
-    product4 = Product("55\" QLED 4K", "Фоновая подсветка", 123000.0, 7)
-    category2 = Category("Телевизоры",
-                         "Современный телевизор, который позволяет наслаждаться просмотром, станет вашим другом и помощником",
-                         [product4])
-
-    print(category2.name)
-    print(category2.description)
-    print(len(category2.products))
-    print(category2.products)
-
-    print(Category.category_count)
-    print(Category.product_count)
