+++ src/smartphone.py
@@ -1,0 +1,28 @@
+from src.product import Product
+
+class Smartphone(Product):
+    """Класс категории смартфоны наследуемый от Product"""
+
+    efficiency: float
+    model: str
+    memory: int
+    color: str
+
+    def __init__(
+        self,
+        name: str,
+        description: str,
+        price: float,
+        quantity: int,
+        efficiency: float,
+        model: str,
+        memory: int,
+        color: str,
+    ):
+        """Инициализация класса смартфонов наследник Product"""
+
+        super().__init__(name, description, price, quantity)
+        self.efficiency = efficiency
+        self.model = model
+        self.memory = memory
+        self.color = color
