+++ tests/test_lawngrass.py
@@ -1,0 +1,11 @@
+
+
+def test_lawngrass_init(lawngrass1):
+    """Инициализация класса наследника газонная трава"""
+    assert lawngrass1.name == "Газонная трава"
+    assert lawngrass1.description == "Элитная трава для газона"
+    assert lawngrass1.price == 500.0
+    assert lawngrass1.quantity == 20
+    assert lawngrass1.country == "Россия"
+    assert lawngrass1.germination_period == "7 дней"
+    assert lawngrass1.color == "Зеленый"
